crawl_service/requirements.txt
fastapi
uvicorn
selenium
psycopg2
python-dotenv
requests
